+++ src/main/java/com/garbagemule/MobArena/commands/admin/EnableCommand.java
@@ -7,6 +7,11 @@ import com.garbagemule.MobArena.framework.Arena;
 import com.garbagemule.MobArena.framework.ArenaMaster;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @CommandInfo(
     name    = "enable",
@@ -49,5 +54,21 @@ public class EnableCommand implements Command
         arena.setEnabled(true);
         arena.getPlugin().saveConfig();
         arena.getGlobalMessenger().tell(sender, "Arena '" + arena.configName() + "' " + ChatColor.GREEN + "enabled");
+    }
+
+    @Override
+    public List<String> tab(ArenaMaster am, Player player, String... args) {
+        if (args.length > 1) {
+            return Collections.emptyList();
+        }
+
+        String prefix = args[0].toLowerCase();
+
+        List<Arena> arenas = am.getArenas();
+
+        return arenas.stream()
+            .filter(arena -> arena.configName().toLowerCase().startsWith(prefix))
+            .map(Arena::configName)
+            .collect(Collectors.toList());
     }
 }