+++ src/main/java/com/garbagemule/MobArena/commands/setup/RemoveArenaCommand.java
@@ -5,6 +5,11 @@ import com.garbagemule.MobArena.commands.CommandInfo;
 import com.garbagemule.MobArena.framework.Arena;
 import com.garbagemule.MobArena.framework.ArenaMaster;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @CommandInfo(
     name    = "removearena",
@@ -33,5 +38,21 @@ public class RemoveArenaCommand implements Command
         am.removeArenaNode(arena);
         am.getGlobalMessenger().tell(sender, "Arena '" + arena.configName() + "' deleted.");
         return true;
+    }
+
+    @Override
+    public List<String> tab(ArenaMaster am, Player player, String... args) {
+        if (args.length > 1) {
+            return Collections.emptyList();
+        }
+
+        String prefix = args[0].toLowerCase();
+
+        List<Arena> arenas = am.getArenas();
+
+        return arenas.stream()
+            .filter(arena -> arena.configName().toLowerCase().startsWith(prefix))
+            .map(Arena::configName)
+            .collect(Collectors.toList());
     }
 }